minor bug in uci parsing...now plays nice with pychess

+++ sillychess/src/sillychess.h
@@ -173,8 +173,8 @@ typedef struct {
 	int quit;
 	int stopped;
 
-	float failHigh;
-	float failHighFirst;
+	int failHigh;
+	int failHighFirst;
 	int nullCut;
 	int lmr;
 	int lmr2;
